added method for calculate line items

+++ src/test/java/com/galvanize/orion/invoicify/InvoiceHelper/InvoiceTestHelper.java
@@ -4,6 +4,7 @@ import com.galvanize.orion.invoicify.entities.Invoice;
 import com.galvanize.orion.invoicify.entities.LineItem;
 import com.galvanize.orion.invoicify.utilities.StatusEnum;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
@@ -12,15 +13,15 @@ import java.util.UUID;
 public class InvoiceTestHelper {
 
     public static LineItem getLineItem() {
-        return LineItem.builder().description("project 1").quantity(10).rate(5.4).build();
+        return LineItem.builder().description("project 1").quantity(10).rate(BigDecimal.valueOf(5.4)).build();
     }
 
     public static LineItem getLineItem2() {
-        return LineItem.builder().description("project 2").quantity(10).rate(4.6).build();
+        return LineItem.builder().description("project 2").quantity(10).rate(BigDecimal.valueOf(4.6)).build();
     }
 
     public static LineItem getLineItem3() {
-        return LineItem.builder().description("project 3").quantity(10).rate(5.0).build();
+        return LineItem.builder().description("project 3").quantity(10).rate(BigDecimal.valueOf(5.0)).build();
     }
 
     public static List<LineItem> getLineItemListWithTwoLineItem() {
